init: tbb
