+++ tasks/tbb/kharin_m_multidimensional_integral_calc/func_tests/main.cpp
@@ -7,7 +7,7 @@
 #include <vector>
 
 #include "core/task/include/task.hpp"
-#include "omp/kharin_m_multidimensional_integral_calc/include/ops_omp.hpp"
+#include "tbb/kharin_m_multidimensional_integral_calc/include/ops_tbb.hpp"
 
 // Тест для 2D интеграла (3x3 сетка)
 TEST(kharin_m_multidimensional_integral_calc_omp, test_integral_3x3) {
