+++ tasks/tbb/kharin_m_multidimensional_integral_calc/include/ops_tbb.hpp
@@ -1,7 +1,5 @@
 #pragma once
 
-#include <omp.h>
-
 #include <cstddef>
 #include <utility>
 #include <vector>
